Works with Sutter equipment
Will test with ASI Tiger Controller next...

Also added the logic necessary to round the stage values before reporting them. I was getting floats with ~10 decimal places...
