+++ src/navigate/model/devices/filter_wheel/ludl.py
@@ -41,6 +41,7 @@ import serial
 # Local Imports
 from navigate.model.devices.filter_wheel.base import FilterWheelBase
 from navigate.model.devices.device_types import SerialDevice
+from navigate.model.devices import MonitoredSerial
 
 # Logger Setup
 p = __name__.split(".")[1]
@@ -116,7 +117,7 @@ class LUDLFilterWheel(FilterWheelBase, SerialDevice):
         """
         logging.debug(f"LUDLFilterWheel - Opening Serial Port {comport}")
         try:
-            return serial.Serial(
+            return MonitoredSerial(
                 comport,
                 baudrate,
                 parity=serial.PARITY_NONE,
